+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/ext/AtomicFile.kt
@@ -47,7 +47,8 @@ fun AtomicFile.writeSnapshot(
  */
 fun AtomicFile.readSnapshotItem(
     engine: Engine,
-    serializer: SnapshotSerializer = SnapshotSerializer()
+    restoreSessionId: Boolean,
+    serializer: SnapshotSerializer = SnapshotSerializer(restoreSessionId)
 ): SessionManager.Snapshot.Item? {
     return readAndDeserialize { json ->
         serializer.itemFromJSON(engine, JSONObject(json))