+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/ext/AtomicFileKtTest.kt
@@ -153,7 +153,7 @@ class AtomicFileKtTest {
 
         file.writeSnapshotItem(item)
 
-        val restoredItem = file.readSnapshotItem(engine)
+        val restoredItem = file.readSnapshotItem(engine, restoreSessionId = true)
         assertNotNull(restoredItem!!)
         assertNotNull(restoredItem.engineSessionState!!)