+++ mobile/android/android-components/components/feature/tab-collections/src/main/java/mozilla/components/feature/tab/collections/Tab.kt
@@ -6,6 +6,7 @@ package mozilla.components.feature.tab.collections
 
 import android.content.Context
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.session.Session
 import mozilla.components.concept.engine.Engine
 
 /**
@@ -29,6 +30,15 @@ interface Tab {
 
     /**
      * Restores a single tab from this collection and returns a matching [SessionManager.Snapshot].
+     *
+     * @param restoreSessionId If true the original [Session.id] of [Session]s will be restored. Otherwise a new ID
+     * will be generated. An app may prefer to use a new ID if it expects sessions to get restored multiple times -
+     * otherwise breaking the promise of a unique ID per [Session].
      */
-    fun restore(context: Context, engine: Engine, tab: Tab): SessionManager.Snapshot
+    fun restore(
+        context: Context,
+        engine: Engine,
+        tab: Tab,
+        restoreSessionId: Boolean = false
+    ): SessionManager.Snapshot
 }